Move copy validation logic into its own file
