+++ cmd/copyValidation.go
@@ -1,0 +1,159 @@
+package cmd
+
+import (
+	"errors"
+	"fmt"
+	"github.com/Azure/azure-storage-azcopy/v10/common"
+	"net/url"
+	"runtime"
+	"strings"
+)
+
+func validateForceIfReadOnly(toForce bool, fromTo common.FromTo) error {
+	targetIsFiles := fromTo.To() == common.ELocation.File() ||
+		fromTo == common.EFromTo.FileTrash()
+	targetIsWindowsFS := fromTo.To() == common.ELocation.Local() &&
+		runtime.GOOS == "windows"
+	targetIsOK := targetIsFiles || targetIsWindowsFS
+	if toForce && !targetIsOK {
+		return errors.New("force-if-read-only is only supported when the target is Azure Files or a Windows file system")
+	}
+	return nil
+}
+func validatePreserveSMBPropertyOption(toPreserve bool, fromTo common.FromTo, overwrite *common.OverwriteOption, flagName string) error {
+	if toPreserve && flagName == PreservePermissionsFlag && (fromTo == common.EFromTo.BlobBlob() || fromTo == common.EFromTo.BlobFSBlob() || fromTo == common.EFromTo.BlobBlobFS() || fromTo == common.EFromTo.BlobFSBlobFS()) {
+		// the user probably knows what they're doing if they're trying to persist permissions between blob-type endpoints.
+		return nil
+	} else if toPreserve && !(fromTo == common.EFromTo.LocalFile() ||
+		fromTo == common.EFromTo.FileLocal() ||
+		fromTo == common.EFromTo.FileFile()) {
+		return fmt.Errorf("%s is set but the job is not between %s-aware resources", flagName, common.Iff(flagName == PreservePermissionsFlag, "permission", "SMB"))
+	}
+
+	if toPreserve && (fromTo.IsUpload() || fromTo.IsDownload()) &&
+		runtime.GOOS != "windows" && runtime.GOOS != "linux" {
+		return fmt.Errorf("%s is set but persistence for up/downloads is supported only in Windows and Linux", flagName)
+	}
+
+	return nil
+}
+
+func validatePreserveOwner(preserve bool, fromTo common.FromTo) error {
+	if fromTo.IsDownload() {
+		return nil // it can be used in downloads
+	}
+	if preserve != common.PreserveOwnerDefault {
+		return fmt.Errorf("flag --%s can only be used on downloads", common.PreserveOwnerFlagName)
+	}
+	return nil
+}
+
+func validateSymlinkHandlingMode(symlinkHandling common.SymlinkHandlingType, fromTo common.FromTo) error {
+	if symlinkHandling.Preserve() {
+		switch fromTo {
+		case common.EFromTo.LocalBlob(), common.EFromTo.BlobLocal(), common.EFromTo.BlobFSLocal(), common.EFromTo.LocalBlobFS():
+			return nil // Fine on all OSes that support symlink via the OS package. (Win, MacOS, and Linux do, and that's what we officially support.)
+		case common.EFromTo.BlobBlob(), common.EFromTo.BlobFSBlobFS(), common.EFromTo.BlobBlobFS(), common.EFromTo.BlobFSBlob():
+			return nil // Blob->Blob doesn't involve any local requirements
+		default:
+			return fmt.Errorf("flag --%s can only be used on Blob<->Blob or Local<->Blob", common.PreserveSymlinkFlagName)
+		}
+	}
+
+	return nil // other older symlink handling modes can work on all OSes
+}
+
+func validateBackupMode(backupMode bool, fromTo common.FromTo) error {
+	if !backupMode {
+		return nil
+	}
+	if runtime.GOOS != "windows" {
+		return errors.New(common.BackupModeFlagName + " mode is only supported on Windows")
+	}
+	if fromTo.IsUpload() || fromTo.IsDownload() {
+		return nil
+	} else {
+		return errors.New(common.BackupModeFlagName + " mode is only supported for uploads and downloads")
+	}
+}
+
+func validatePutMd5(putMd5 bool, fromTo common.FromTo) error {
+	// In case of S2S transfers, log info message to inform the users that MD5 check doesn't work for S2S Transfers.
+	// This is because we cannot calculate MD5 hash of the data stored at a remote locations.
+	if putMd5 && fromTo.IsS2S() {
+		glcm.Info(" --put-md5 flag to check data consistency between source and destination is not applicable for S2S Transfers (i.e. When both the source and the destination are remote). AzCopy cannot compute MD5 hash of data stored at remote location.")
+	}
+	return nil
+}
+
+func validateMd5Option(option common.HashValidationOption, fromTo common.FromTo) error {
+	hasMd5Validation := option != common.DefaultHashValidationOption
+	if hasMd5Validation && !fromTo.IsDownload() {
+		return fmt.Errorf("check-md5 is set but the job is not a download")
+	}
+	return nil
+}
+
+// Valid tag key and value characters include:
+// 1. Lowercase and uppercase letters (a-z, A-Z)
+// 2. Digits (0-9)
+// 3. A space ( )
+// 4. Plus (+), minus (-), period (.), solidus (/), colon (:), equals (=), and underscore (_)
+func isValidBlobTagsKeyValue(keyVal string) bool {
+	for _, c := range keyVal {
+		if !((c >= '0' && c <= '9') || (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z') || c == ' ' || c == '+' ||
+			c == '-' || c == '.' || c == '/' || c == ':' || c == '=' || c == '_') {
+			return false
+		}
+	}
+	return true
+}
+
+// ValidateBlobTagsKeyValue
+// The tag set may contain at most 10 tags. Tag keys and values are case sensitive.
+// Tag keys must be between 1 and 128 characters, and tag values must be between 0 and 256 characters.
+func validateBlobTagsKeyValue(bt common.BlobTags) error {
+	if len(bt) > 10 {
+		return errors.New("at-most 10 tags can be associated with a blob")
+	}
+	for k, v := range bt {
+		key, err := url.QueryUnescape(k)
+		if err != nil {
+			return err
+		}
+		value, err := url.QueryUnescape(v)
+		if err != nil {
+			return err
+		}
+
+		if key == "" || len(key) > 128 || len(value) > 256 {
+			return errors.New("tag keys must be between 1 and 128 characters, and tag values must be between 0 and 256 characters")
+		}
+
+		if !isValidBlobTagsKeyValue(key) {
+			return errors.New("incorrect character set used in key: " + k)
+		}
+
+		if !isValidBlobTagsKeyValue(value) {
+			return errors.New("incorrect character set used in value: " + v)
+		}
+	}
+	return nil
+}
+
+func validateMetadataString(metadata string) error {
+	if strings.EqualFold(metadata, common.MetadataAndBlobTagsClearFlag) {
+		return nil
+	}
+	metadataMap, err := common.StringToMetadata(metadata)
+	if err != nil {
+		return err
+	}
+	for k := range metadataMap {
+		if strings.ContainsAny(k, " !#$%^&*,<>{}|\\:.()+'\"?/") {
+			return fmt.Errorf("invalid metadata key value '%s': can't have spaces or special characters", k)
+		}
+	}
+
+	return nil
+}
